requirements.txt
translate
pillow